requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
python-jose==3.3.0
pyasn1==0.6.1
rsa==4.9
passlib==1.7.4
anthropic==0.43.0
# google-genai removed — using Gemini REST API directly (avoids grpcio SIGILL on Apple Silicon)
python-multipart==0.0.18
psycopg[binary]==3.2.4
celery[redis]==5.4.0
redis==5.2.1
dnspython==2.7.0
python-whois==0.9.4
tldextract==5.1.3
python-Levenshtein==0.26.1
homoglyphs==2.0.4
Pillow==11.1.0
aiohttp==3.11.11
requests==2.32.3
google-play-scraper==1.2.7
playwright==1.49.1
minio==7.2.12
reportlab==4.2.5
